backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
yt-dlp>=2023.3.4
pyannote.audio>=2.1.1
openai-whisper>=20230314
numpy>=1.24.2
torch>=2.0.0
torchaudio>=2.0.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
